Move GLSpectrum to GLSpectrumView and create new GLSpectrum that instances GLSpectrumView and SpectrumMeasurements
