+++ sdrgui/mainspectrum/mainspectrumgui.cpp
@@ -26,15 +26,13 @@
 
 #include "mainwindow.h"
 #include "gui/glspectrum.h"
-#include "gui/glspectrumtop.h"
 #include "gui/glspectrumgui.h"
 #include "gui/workspaceselectiondialog.h"
 #include "dsp/spectrumvis.h"
 #include "mainspectrumgui.h"
 
-MainSpectrumGUI::MainSpectrumGUI(GLSpectrumTop *spectrumTop, GLSpectrum *spectrum, GLSpectrumGUI *spectrumGUI, QWidget *parent) :
+MainSpectrumGUI::MainSpectrumGUI(GLSpectrum *spectrum, GLSpectrumGUI *spectrumGUI, QWidget *parent) :
     QMdiSubWindow(parent),
-    m_spectrumTop(spectrumTop),
     m_spectrum(spectrum),
     m_spectrumGUI(spectrumGUI),
     m_deviceType(DeviceRx),
@@ -115,7 +113,7 @@ MainSpectrumGUI::MainSpectrumGUI(GLSpectrumTop *spectrumTop, GLSpectrum *spectru
     m_topLayout->addWidget(m_hideButton);
 
     m_spectrumLayout = new QHBoxLayout();
-    m_spectrumLayout->addWidget(spectrumTop);
+    m_spectrumLayout->addWidget(spectrum);
     spectrum->setSizePolicy(QSizePolicy::MinimumExpanding, QSizePolicy::MinimumExpanding);
     m_spectrumGUILayout = new QHBoxLayout();
     m_spectrumGUILayout->addWidget(spectrumGUI);
@@ -143,7 +141,7 @@ MainSpectrumGUI::MainSpectrumGUI(GLSpectrumTop *spectrumTop, GLSpectrum *spectru
     connect(this, SIGNAL(forceShrink()), this, SLOT(shrinkWindow()));
     connect(m_hideButton, SIGNAL(clicked()), this, SLOT(hide()));
 
-    connect(spectrum, &GLSpectrum::requestCenterFrequency, this, &MainSpectrumGUI::onRequestCenterFrequency);
+    connect(spectrum->getSpectrumView(), &GLSpectrumView::requestCenterFrequency, this, &MainSpectrumGUI::onRequestCenterFrequency);
     connect(spectrumGUI, &GLSpectrumGUI::requestCenterFrequency, this, &MainSpectrumGUI::onRequestCenterFrequency);
 
     m_resizer.enableChildMouseTracking();
@@ -153,7 +151,7 @@ MainSpectrumGUI::MainSpectrumGUI(GLSpectrumTop *spectrumTop, GLSpectrum *spectru
 MainSpectrumGUI::~MainSpectrumGUI()
 {
     qDebug("MainSpectrumGUI::~MainSpectrumGUI");
-    m_spectrumLayout->removeWidget(m_spectrumTop);
+    m_spectrumLayout->removeWidget(m_spectrum);
     m_spectrumGUILayout->removeWidget(m_spectrumGUI);
     delete m_sizeGripBottomRight;
     delete m_bottomLayout;
